refactor(backup): move event based hold code to dedicated files
At first, I thought that the whole event based hold implementation
would fit into worker_retention_lock file, but it would need
to contain both eventBasedHoldHandler and additional machinery
needed for cleaning holds from vanished dirs, which is a little
too much.

This commit simply moves what can be moved without any changes
to dedicated worker_event_based_hold, and the next commits will
aim have fully separate execution paths for different modes,
as the vanished dir cleanup changes the manifest cleanup quite a lot.
